Non exported functions can be found here.

LimnoSES.type2dict — Function

type2dict(struct; prefix = "")

Converts a struct into a Dict. Borrowed from DrWatson, but extended to allow some prefix to be attached. This is helpful when merging multiple structs into one Dict.

LimnoSES.nutrient_load! — Function

nutrient_load(model::ABM, series<:NutrientSeries)

Updates lake nutrient concentration accorting to series type.

See NutrientSeries for more details.

LimnoSES.active_interventions — Function

active_interventions(municipality, year)

Returns all active interventions in the planner. Due to the year = -1 -> always active convention we must merge the current year's plan with the -1 key (if extant).