Non exported functions can be found here.

LimnoSES.type2dict — Function

type2dict(struct; prefix = "")

Converts a struct into a Dict. Borrowed from DrWatson, but extended to allow some prefix to be attached. This is helpful when merging multiple structs into one Dict.

LimnoSES.nutrient_load! — Function

nutrient_load(model::ABM, series<:NutrientSeries)

Updates lake nutrient concentration accorting to series type.

See NutrientSeries for more details.

LimnoSES.active_interventions — Function

active_interventions(municipality, year)

Returns all active interventions in the planner. Due to the year = -1 -> always active convention we must merge the current year's plan with the -1 key (if extant).

Decisions Private Functions

LimnoSES.create_test_model — Function

create_test_model(model)

Creates a complete copy of the current model, with a modified set of interventions. All "completed" interventions (i.e. ones that have happened in the models' past) are ignored. This test model is then used in the optimisation procedure.

LimnoSES.apply_policies! — Function

apply_policies!(x, test::ABM)

Applies new test values to all active intervention properties. Is used before each optimisation call, but also to finalise the decision process via a best_candidate call.

LimnoSES.update_true_model! — Function

update_true_model!(test, model)

Reconstructs each municipalities interventions property in the model after a successful optimisation run. Be careful with the model order here, as there's no simple way to differentiate the two of them via types.

LimnoSES.cost — Function

cost(x, u0, p, test::ABM)

Optimization function. To be used in in conjunction with the bboptimize call only. Needs to be overloaded slightly, since we want to reset the lake dynamics at each call, so model.lake.u and model.lake.p are expected to be passed into the second and third variables respectively, with the cut down "test" version of the model being the last value.

LimnoSES.weightedfitness — Function

weightedfitness(f, weights)

Handle objective weights. Should be used as an aggregator in bboptimize.