None of this is final, but some current public facing functions:

LimnoSES.initialise — Function

initialise()

Returns a populated model initialised and ready to run.

LimnoSES.planner — Function

planner(plan(Angling))
planner(plan(Planting; rate=5e-3),
        plan(Trawling, 1:3))

Provides a complete schedule of interventions for a Municipality. Must be used in conjunction with plan.

LimnoSES.plan — Function

plan(Angling) # Assume always on
plan(Angling; rate = 2.5e-3) # Always on with custom rate
plan(Angling, 7; rate = 3.2e-3) # Only one year (custom rate)
plan(Angling, 3:5) # Only years 3 to 5
plan(Angling, [(period = 1:4, ),
               (year = 5, rate = 7.2e-3),
               (period = 7:9, )]) # Active in years 1-5, 7-9 with a custom rate
                                  # in year 5

Helper that provides complex scheduling for interventions with a simple interface.

The return type is Dict{Int, Vector{Intervention}}, were the key is each year the collection of interventions will be active. As a convention, year -1 denotes an 'always active' intervention.

The result can be provided to municipality.interventions, although this function should almost always used in conjuction with planner.

NutrientSeries

NutrientSeries is an abstract type from which concrete types can be implemented to describe the dynamics of nutrient introduction to the lake.

LimnoSES.Constant — Type

Constant()

Nutrient level remains constant at the level of init_nutrients.

LimnoSES.Dynamic — Type

Dynamic()

Nutrient runoff is managed by the municipality by incentivising households to upgrade sewage systems that seep P into the lake.

LimnoSES.TransientUp — Type

TransientUp(;start_year = 11, post_target_series = Constant())

Synthetic nutrient profile that alters lake dynamics regardless of municipal management.

start_year: year when nutrients begin to increase with a rate of nutrient_change.

post_target_series: behaviour after target_nutrients value is reached. Default is TransientDown(start_year = 0, post_target_series = Constant())

Warning

Post target series selection must include a final Constant phase, otherwise an infinite recursion cascade will occur.

LimnoSES.TransientDown — Type

TransientDown(;start_year = 11, post_target_series = Constant())

Synthetic nutrient profile that alters lake dynamics regardless of municipal management.

start_year: year when nutrients begin to decrease with a rate of nutrient_change.

post_target_series: behaviour after target_nutrients value is reached. Default is TransientUp(start_year = 0, post_target_series = Constant())

Warning

Post target series selection must include a final Constant phase, otherwise an infinite recursion cascade will occur.